+++ src/components/widgets/spinners/LeftSpinner.js
@@ -1,13 +1,13 @@
 import CircularProgress from "@material-ui/core/CircularProgress";
 import React from "react";
 import { makeStyles, withStyles } from '@material-ui/core/styles';
-import colors from "../../styles/colors";
+import colors from "../../../styles/colors";
 
 const useStyles = makeStyles(theme => ({
   root: {
     display: 'flex',
     flex: 1,
-    justifyContent: 'center',
+    justifyContent: 'left',
     alignItems: 'center',
     flexDirection: 'row',
   },
@@ -15,11 +15,11 @@ const useStyles = makeStyles(theme => ({
 
 const ColorCircularProgress = withStyles({
   root: {
-    color: colors.presqtBlue
+    color: colors.presqtBlue,
   },
 })(CircularProgress);
 
-export default function Spinner() {
+export default function LeftSpinner() {
   const classes = useStyles();
 
   return (
