+++ src/components/widgets/spinners/Spinner.js
@@ -1,13 +1,13 @@
 import CircularProgress from "@material-ui/core/CircularProgress";
 import React from "react";
 import { makeStyles, withStyles } from '@material-ui/core/styles';
-import colors from "../../styles/colors";
+import colors from "../../../styles/colors";
 
 const useStyles = makeStyles(theme => ({
   root: {
     display: 'flex',
     flex: 1,
-    justifyContent: 'left',
+    justifyContent: 'center',
     alignItems: 'center',
     flexDirection: 'row',
   },
@@ -15,11 +15,11 @@ const useStyles = makeStyles(theme => ({
 
 const ColorCircularProgress = withStyles({
   root: {
-    color: colors.presqtBlue,
+    color: colors.presqtBlue
   },
 })(CircularProgress);
 
-export default function LeftSpinner() {
+export default function Spinner() {
   const classes = useStyles();
 
   return (
